third_party/blink/web_tests/http/tests/websocket/handshake-fail-by-mismatch-protocol-header-expected.txt

Test whether WebSocket handshake fails if the server sends mismatched Sec-WebSocket-Protocol header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS onerror() was called
PASS closeEvent.wasClean is false
PASS onerror() was called
PASS closeEvent.wasClean is false
PASS onerror() was called
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
